Remove NO_PYTHON in tracing
