+++ torch/csrc/Device.cpp
@@ -7,6 +7,7 @@
 #include "torch/csrc/utils/object_ptr.h"
 #include "torch/csrc/utils/python_arg_parser.h"
 #include "torch/csrc/utils/python_strings.h"
+#include "torch/csrc/utils/pybind.h"
 
 PyObject *THPDevice_New(const torch::Device& device)
 {
