+++ torch/csrc/jit/tracer.h
@@ -8,9 +8,6 @@
 #include "torch/csrc/autograd/function_hook.h"
 #include "torch/csrc/autograd/variable.h"
 #include "torch/csrc/utils/auto_unique_ptr.h"
-#ifndef NO_PYTHON
-#include "torch/csrc/utils/pybind.h"
-#endif
 #include <memory>
 #include <mutex>
 #include <vector>
@@ -22,10 +19,6 @@ namespace torch { namespace jit { namespace tracer {
 
 using torch::autograd::Variable;
 using variable_list = std::vector<Variable>;
-
-#ifndef NO_PYTHON
-std::string getPythonInterpreterStackTrace();
-#endif
 
 namespace detail {
 
@@ -281,18 +274,36 @@ struct PreTraceInfo {
 };
 
 PreTraceInfo preRecordTrace(Symbol op, at::ArrayRef<Variable> inputs);
-#ifndef NO_PYTHON
-PreTraceInfo preRecordPythonTrace(
-    THPObjectPtr pyobj, std::string arg_types, at::ArrayRef<Variable> inputs,
-    pyobj_list scalar_args);
-
-std::shared_ptr<Graph> createGraphByTracing(
-        py::function func,
-        variable_list inputs,
-        size_t num_inputs);
-#endif
 void postRecordTrace(const PreTraceInfo& info, at::ArrayRef<Variable> outputs);
 
 autograd::Variable getSizeOf(const autograd::Variable& var, int64_t dim);
+
+void recordSourceLocation(Node* n);
+void setRecordSourceLocation(void (*v)(Node*));
+
+// We must record the nodes of inputs before we actually carry out
+// the operation, because an inplace operation may destroy the information
+// we're interested in.  See #4480.
+template<typename F>
+PreTraceInfo makePreTraceInfo(at::ArrayRef<Variable> inputs, F ctor) {
+  PreTraceInfo info;
+  info.state = getTracingState(inputs);
+  auto& graph = info.state->graph;
+  auto state_lock = info.state->lock();
+
+  Node *n = ctor(info.state, *graph);
+  recordSourceLocation(n);
+
+  for (Variable input : inputs) {
+    n->addInput(getValueTrace(info.state, input));
+  }
+
+  // NB: Order matters. This must append after inputs but before outputs.
+  graph->appendNode(n);
+
+  info.n = n;
+
+  return info;
+}
 
 }}} // namespace torch::jit::tracer
